Feature/equations

+++ services/userService.js
@@ -1,8 +1,10 @@
 const config = require('../_helpers/config.json');
 const jwt = require('jsonwebtoken');
 const bcrypt = require('bcryptjs');
+const roleService = require('./roleService');
 
 const User = require('../models/User');
+const Roles = require('../models/UserRole');
 
 async function authenticate({ email, password }) {
     const user = await User.findOne({ email });
@@ -15,6 +17,9 @@ async function authenticate({ email, password }) {
             ...userWithoutHash,
             token
         };
+    }
+    if (!user) {
+        throw `email or password is incorrect`;
     }
 }
 
@@ -41,10 +46,16 @@ async function create(userParam) {
 
     // save user
     const createdUser = await user.save();
-    delete createdUser.password;
+    const _user = createdUser.toObject();
+
+    if (!await Roles.findOne({userId: _user._id})) {
+        const newRole = new Roles({userId: _user._id});
+        await newRole.save();
+    }
+    delete _user.password;
     return {
 		message: 'User created successfully ...',
-		user: createdUser
+		user: _user
 	};
 }
 
@@ -56,10 +67,13 @@ async function update(id, userParam) {
     if (user.email !== userParam.email && await User.findOne({ email: userParam.email })) {
         throw 'Email "' + userParam.email + '" is already taken';
     }
-
+    const role = await roleService.findUserRole(id);
+    if (!role.isAdmin) {
+        throw `You Don't Have Admin Access`;
+    }
     // hash password if it was entered
     if (userParam.password) {
-        userParam.hash = bcrypt.hashSync(userParam.password, 10);
+        userParam.password = bcrypt.hashSync(userParam.password, 10);
     }
 
     // copy userParam properties to user
@@ -73,6 +87,10 @@ async function update(id, userParam) {
 }
 
 async function _delete(id) {
+    const role = await roleService.findUserRole(id);
+    if (!role.isAdmin) {
+        throw `You Don't Have Admin Access`;
+    }
     await User.findByIdAndRemove(id);
 }
